requirements.txt
amqp==2.4.2
argon2-cffi==19.1.0
arrow==0.13.1
astroid==2.2.5
autopep8==1.4.4
Babel==2.7.0
backports.csv==1.0.7
bcrypt==3.1.6
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
Click==7.0
colorama==0.4.1
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.0.6
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3==11.0.0
django-braces==1.13.0
django-celery-beat==1.4.0
django-celery-results==1.0.4
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-debug-toolbar==2.0
django-embed-video==1.2.0
django-import-export==1.2.0
django-js-asset==1.2.2
django-mathfilters==0.4.0
django-model-utils==3.1.2
django-pagedown==1.0.6
django-phonenumber-field==3.0.1
django-sslserver==0.20
django-star-ratings==0.7.0
django-timezone-field==3.0
docopt==0.6.2
et-xmlfile==1.0.1
ez-setup==0.9
Flask==1.0.2
gevent==1.4.0
greenlet==0.4.15
html5lib==1.0.1
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4
Jinja2==2.10.1
kombu==4.5.0
lazy-object-proxy==1.4.1
locustio==0.11.0
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.1
num2words==0.5.10
numpy==1.16.3
oauthlib==3.0.1
odfpy==1.4.0
openpyxl==2.6.2
pandas==0.24.2
phonenumbers==8.10.14
Pillow==5.4.1
psycopg2==2.8.2
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
pylint==2.3.1
pylint-django==2.0.11
pylint-plugin-utils==0.5
PyPDF2==1.26.0
python-crontab==2.3.6
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2018.9
PyYAML==5.1
pyzmq==18.0.1
qrcode==6.1
redis==3.2.1
reportlab==3.5.17
requests==2.21.0
requests-oauthlib==1.2.0
rope==0.14.0
selenium==3.141.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
sorl-thumbnail==12.5.0
sqlparse==0.3.0
swapper==1.1.0
tablib==0.13.0
typed-ast==1.4.0
urllib3==1.24.1
vine==1.3.0
webencodings==0.5.1
Werkzeug==0.15.2
wrapt==1.11.2
xhtml2pdf==0.2.3
xlrd==1.2.0
xlwt==1.3.0
